changed db to mariadb, some refactor

+++ src/main/java/web/model/UserModel.java
@@ -9,7 +9,7 @@ public class UserModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private long id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "surname")
@@ -75,7 +75,7 @@ public class UserModel {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
